fix: (loop) hafnian for odd-size and null matrices
Instead of erroring out, define the hafnian and loop hafnian also on
odd-sized square matrices and on 0x0 matrices, as 0 and 1 respectively.
Also use `eltype` when constructing all-zero matrices within the
functions, so that the types are consistent through the calculations and
there are no `InexactError`s when trying to overwrite the matrix
elements.