test: fix MPS check of squeezed product state
Do not check the MPS construction of a product of squeezed states by
looking at the individual MPS matrices: they do not necessarily coincide
with the coefficients of the state for each vector of the number basis
(for each mode). Check instead the coefficients of the whole product
state, i.e. the whole MPS: individual matrices could get modified in
unexpected ways by "MPS gauge transformation", which cancel out when the
whole state is considered.