test: compare only absolute value of coefficients
The "direct" Franck-Condon formula and the full MPS-building algorithm,
albeit starting from the same initial covariance matrix, follow
different paths to calculate the coefficient of the state on the number
eigenbasis, and compute the decomposition of different intermediate
matrices. These paths are equivalent in the sense that the final state
is the same, but there may be an overall leftover phase factor.
This of course is not a problem since in the end the results represent
the same physical state, but it means that a straight comparison
of the coefficients obtained with the two methods will yield a failed
test, since the phases in general won't match.
Without a way to compare two vectors up to a phase, only the absolute
values of the coefficients will be compared for now.