feat: full-MPS simulation with attenuator channel
Add some utilities to manage matrix-product states in the superboson
formalism, so that a full simulation with MPSs from the beginning to
the end can be performed.
Define a new ITensor operator that implements the attenuator (pure loss)
channel.
Test the new functions in a small new test set, checking the result of
applying the new attenuator channel against some analytical formulas.