refactor: relax check if 1st moments aren't real
Warn if 1st moments from an MPS are not real only if the overall norm of
the first moments vector is above a certain threshold.
It's not uncommon that the first moments are zero, and in this case
calling `isapprox` with a zero as argument always returns `false`,
and the warning is triggered even if it's not necessary; the new
additional condition prevents this.