fix: cast MPS probabilities to real numbers
In the sampling algorithm, the probabilities associated to each basis
state at each step (i.e. the diagonal elements of the density matrix of
the reduced state), are in theory real numbers, but in reality they are
complex since all intermediate operations are performed with complex
numbers and there's a small imaginary part leftover from numerical
imprecisions. To solve this, after checking that the imaginary part is
indeed satisfyingly small, we just throw it away.