refactor: collect samples in `UInt8` matrix
Collect the sampled states in the `sample_displaced` method in a big
matrix of `UInt8`s instead of a list of `Int` vectors, so that it takes
up less space and it is faster to create. It's very unlikely that the
coefficients of the individual samples exceed 255 anyway, at least for
now.