feat: new `SuperBosonMPS` type
Define a new struct `SuperBosonMPS`, subtype of ITensorMPS's
`AbstractMPS`, that represents and MPS describing a mixed many-boson
state in the superboson formalism.
Its inner structure is identical to the one of an ordinary MPS, but the
different name allows us to define different `expect`, `sample` (and so
on) functions on it with the same name as ITensorMPS's methods.