feat: new generic 1st and 2nd moments functions
Write new `firstmoments` and `covariancematrix` functions that work
with both standard MPSs and superbosonic ones, with the same code.
This was made possible by mimicking the `ITensorMPS` MPS interface,
defining `expect` and `correlation_matrix` methods for the
`SuperBosonMPS` type so that they could be called in the exact
same way for this type and ITensor's `MPS` type.

With this feature, the `firstmoments` and `covariancematrix` methods can
be used seamlessly on standard MPSs too, without having to define a
separate function.