feat!: exact matrix coeffs of squeezing operator
The squeezing operator `S(z)` calculated as `exp(z/2 a*² - z̄/2 a²)` uses
the truncated versions of the creation and annihilation operators, as
provided by ITensor, and as a consequence its coefficients (in the
eigenbasis of the number operator, for example) depend on the dimension
of the truncated space. This dependence can be very noticeable, such
that even the coefficients of the zero- or one-particle sector, are
significantly off from their correct value unless the squeezing
parameter `z` is such that `|z| << 1`.

In order to remedy this, the operator should be computed with a
sufficiently higher cutoff, then truncated to the desired dimension, but
this higher temporary cutoff would depend on the squeezing parameter,
and it may lead to the computation of the exponential of very large
matrix.

A better approach is to compute the matrix elements of the operator
by hand, skipping the calculation of the exponential altogether.
This way, the coefficients are correct regardless of the chosen cutoff.
We use this new method.

The downside is that `S(z)` is not unitary anymore, and in particular it
is not true anymore that `S(-z) S(z) = I`, even though an approximate
equality is still attainable if `|z| << 1`.
A test that made use of this property was therefore removed.